+++ discoart/runner.py
@@ -173,6 +173,7 @@ def do_run(args, models, device, events) -> 'DocumentArray':
 
         num_step = _MAX_DIFFUSION_STEPS - t_int
         scheduler = _get_current_schedule(schedule_table, num_step)
+        is_cuts_visualized = False
 
         with torch.enable_grad():
 
@@ -265,10 +266,23 @@ def do_run(args, models, device, events) -> 'DocumentArray':
                         IC_Grey_P=scheduler.cut_icgray_p,
                         skip_augs=scheduler.skip_augs,
                     )
-                    clip_in = normalize(cuts(x_in.add(1).div(2)))
+
+                    clip_in = cuts(x_in.add(1).div(2))
+
+                    if args.visualize_cuts and not is_cuts_visualized:
+                        _cuts_da = DocumentArray.empty(clip_in.shape[0])
+                        _cuts_da.tensors = (clip_in * 255).detach().cpu().numpy()
+                        _cuts_da.plot_image_sprites(
+                            os.path.join(output_dir, f'{_nb}-cuts-{num_step}.png'),
+                            show_index=True,
+                            channel_axis=0,
+                        )
+                        is_cuts_visualized = True
 
                     image_embeds = (
-                        model_stat['clip_model'].encode_image(clip_in).unsqueeze(1)
+                        model_stat['clip_model']
+                        .encode_image(normalize(clip_in))
+                        .unsqueeze(1)
                     )
 
                     dists = spherical_dist_loss(
@@ -454,6 +468,7 @@ scheduling tracking, please set `WANDB_MODE=online` before running/importing Dis
 
                 is_save_step = args.save_rate > 0 and j % args.save_rate == 0
                 is_complete = cur_t == -1
+                is_display_step = args.display_rate > 0 and j % args.display_rate == 0
 
                 threads.append(
                     _sample_thread(
@@ -470,6 +485,7 @@ scheduling tracking, please set `WANDB_MODE=online` before running/importing Dis
                         is_save_step or is_complete,
                         args.gif_fps > 0,
                         args.image_output,
+                        is_display_step,
                     )
                 )